Add an integration test for all fixers
