+++ tests/fixes/mod.rs
@@ -1,12 +1,13 @@
-mod quote_character;
-
 use crate::common::TestDir;
 
 mod duplicated_key;
 mod ending_blank_line;
 mod extra_blank_line;
 mod incorrect_delimiter;
+mod key_without_value;
 mod leading_character;
+mod lowercase_key;
+mod quote_character;
 mod space_character;
 mod trailing_whitespace;
 mod unordered_key;
@@ -19,37 +20,6 @@ fn correct_file() {
     testdir.test_command_fix_success(String::new());
 
     assert_eq!(testfile.contents().as_str(), "ABC=DEF\nD=BAR\n\nFOO=BAR\n");
-
-    testdir.close();
-}
-
-#[test]
-fn lowercase_key() {
-    let testdir = TestDir::new();
-    let testfile = testdir.create_testfile(".env", "abc=DEF\n\nfOO=BAR\n");
-    let expected_output = String::from(
-        "Fixed warnings:\n\
-        .env:1 LowercaseKey: The abc key should be in uppercase\n\
-        .env:3 LowercaseKey: The fOO key should be in uppercase\n",
-    );
-    testdir.test_command_fix_success(expected_output);
-
-    assert_eq!(testfile.contents().as_str(), "ABC=DEF\n\nFOO=BAR\n");
-
-    testdir.close();
-}
-
-#[test]
-fn key_without_value() {
-    let testdir = TestDir::new();
-    let testfile = testdir.create_testfile(".env", "FOO\n\nBAR=\n\nBAZ=QUX\n");
-    let expected_output = String::from(
-        "Fixed warnings:\n\
-        .env:1 KeyWithoutValue: The FOO key should be with a value or have an equal sign\n",
-    );
-    testdir.test_command_fix_success(expected_output);
-
-    assert_eq!(testfile.contents().as_str(), "FOO=\n\nBAR=\n\nBAZ=QUX\n");
 
     testdir.close();
 }
